models_api/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
bcrypt==4.1.2
PyJWT==2.8.0
